Added group-by-key.scm & .gitignore
Completed question 3. This question requires a slightly more complex function
however I seem to have managed it.

In the specification it states the requirement:
'The solution performs a sub-quadratic number of allocations'

For each element of the association list passed to the function group-by-key
2 cons (construct) operations are performed however both assoc-all and
remove-assoc-all are used which in turn use cons.

1. assoc-all uses cons at most once per element of the association list.
2. remove-assoc-all uses cons at most once per element of the association list.

This would appear to therefore use (2n^2 + 2n) quadratic uses of cons however
this is not the case as the size of the list decreases in size with every use
in remove-assoc-all.

Two extreme cases:
1. All keys are the same:
	assoc-all cons uses = n*1
	remove-all cons uses = 0
2. All keys are different:
	assoc-all cons uses = 1*n
	remove-all cons uses = (n-1)+(n-2)+(n-3)+...+1 = (n^2)/2

This is quadratic and is therefore not 'sub-quadratic'.
I cannot see how this can be improved unless this is considered 'sub-quadratic'.

Fraser 27/10/2016 20:27:42